Change version to alpha
The current version certainly works and has the basic API baked in but
I wouldn't like to suggest this is in any way indicative of a final 1.0
release in terms of quality. The pluralisation in particular needs a
proper bit of work to address loads of shortcomings. And there's no
tests.